Contain all elements and remove html full height
With wanting to create a semi-responsive design, I found trying to fix
a perfect height for the downloaded image too difficult (or i was too
lazy lol). Thus I removed the html having 100% height and placed all
the HTML elements into a container which I can centre on the page.